src/utils/validators.py
import re
from typing import Any

class Validators:
    @staticmethod
    def is_valid_email(email: str) -> bool:
        """Validate email format."""
        pattern = r'^[\w\.-]+@[\w\.-]+\.\w+$'
        return bool(re.match(pattern, email))
    
    @staticmethod
    def is_valid_phone(phone: str) -> bool:
        """Validate phone number format (simple)."""
        pattern = r'^\+?1?\d{9,15}$'
        return bool(re.match(pattern, phone))
    
    @staticmethod
    def is_within_range(value: float, min_val: float, max_val: float) -> bool:
        """Check if value is within specified range."""
        return min_val <= value <= max_val